+++ packages/core/src/helpers/getSchemaTypeByName.ts
@@ -1,13 +1,5 @@
 import { MarkType, NodeType, Schema } from 'prosemirror-model'
 
 export default function getSchemaTypeByName(name: string, schema: Schema): NodeType | MarkType | null {
-  if (schema.nodes[name]) {
-    return schema.nodes[name]
-  }
-
-  if (schema.marks[name]) {
-    return schema.marks[name]
-  }
-
-  return null
+  return schema.nodes[name] || schema.marks[name] || null
 }