+++ packages/vue-3/src/VueNodeViewRenderer.ts
@@ -92,8 +92,14 @@ class VueNodeView extends NodeView<Component, Editor, VueNodeViewRendererOptions
         provide('onDragStart', onDragStart)
         provide('decorationClasses', this.decorationClasses)
 
-        return (this.component as any).setup?.(props)
+        return (this.component as any).setup?.(props, {
+          expose: () => undefined,
+        })
       },
+      // add support for scoped styles
+      // @ts-ignore
+      // eslint-disable-next-line
+      __scopeId: this.component.__scopeId,
     })
 
     this.renderer = new VueRenderer(extendedComponent, {